% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{co2_emission_2022}
\alias{co2_emission_2022}
\title{CO2 Emissions dataset}
\format{
A data frame with 28 rows and 3 columns
\describe{
\item{region}{eGRID subregion abbreviation}
\item{co2}{CO2 total output emission rate (lb/MWh)}
\item{year}{year of emissions, currently just 2022}
}
}
\source{
https://www.epa.gov/egrid/data-explorer
}
\usage{
co2_emission_2022
}
\description{
Collected from U.S. EPA eGRID Data Explorer
}
\keyword{datasets}
