% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combustion}
\alias{combustion}
\title{Combustion vs non-combustion resource mix data}
\format{
A data frame with 56 rows and 4 columns
\describe{
\item{region}{eGRID subregion abbreviation}
\item{fuel_type}{fuel type(s)}
\item{resource_mix}{percent of total resource mix}
\item{year}{year of data collection. currently only 2023}
}
}
\source{
https://www.epa.gov/egrid/data-explorer
}
\usage{
combustion
}
\description{
Resource Mix by combustible vs. non-combustible fuels
by eGRID subregion, 2023. Data collected from EPA eGRID Data Explorer.
}
\keyword{datasets}
