% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egrid_palettes.R
\docType{data}
\name{egrid_palettes}
\alias{egrid_palettes}
\title{Color Palettes for eGRID}
\format{
\subsection{\code{who}}{

A named list with hex color codes
}
}
\source{
eGRID Data Visualization Style Guide
}
\usage{
egrid_palettes
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
