% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gg_ex_bar}
\alias{gg_ex_bar}
\title{Example barplot of 2023 eGRID resource mix}
\format{
An object of class \code{gg} (inherits from \code{ggplot}) of length 11.
}
\source{
CAPD Data Visualization Style Guide
}
\usage{
gg_ex_bar
}
\description{
ggplot object for use in examples
}
\keyword{datasets}
