% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcapdthemes-package.R
\docType{package}
\name{ggcapdthemes-package}
\alias{ggcapdthemes}
\alias{ggcapdthemes-package}
\title{ggcapdthemes: Custom ggplot2 theme for EPA CAPD visualizations}
\description{
Provides a custom ggplot2 theme and functionality to incorporate guidance from the EPA's Clean Air Power Division (CAPD) Data Visualization Style Guide. Examples show how to implement the theme and color palettes and demonstrate best practices for building various chart types in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/USEPA/ggcapdthemes}
  \item Report bugs at \url{https://github.com/USEPA/ggcapdthemes/issues}
}

}
\author{
\strong{Maintainer}: Johnathan Tafoya \email{Tafoya.Johnathan@epa.gov}

Authors:
\itemize{
  \item Travis Johnson \email{Johnson.Travis@epa.gov}
  \item Marschall Furman \email{Marschall.Furman@abtglobal.com}
}

}
\keyword{internal}
