% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resource_mix}
\alias{resource_mix}
\title{Resource Mix Data}
\format{
A data frame with 11 rows and 6 columns
\describe{
\item{year}{year of data collection}
\item{nameplate_capacity_mw}{Nameplate Capacity (MW)}
\item{net_generation_m_wh}{Net Generation (MWh)}
\item{resource}{cleaned resource name for use in R}
\item{generation_resource_mix}{Generation Resource Mix (percent)}
\item{resource_label}{Original Resource name, for charts}
}
}
\source{
eGRID 2018-2023 Summary data
}
\usage{
resource_mix
}
\description{
Collected from eGRID 2018-2023 Summary datasets Table 2. Example data for demonstrating eGRID themes and scales for barplots
}
\keyword{datasets}
