% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_egrid_discrete}
\alias{scale_color_egrid_discrete}
\title{eGRID discrete color scales}
\usage{
scale_color_egrid_discrete(palette = "fuel_type", reverse = FALSE, ...)
}
\arguments{
\item{palette}{name of palette, from \code{egrid_palettes}. Defaults to "fuel_type".}

\item{reverse}{logical to indicate if palette order should be reversed. Defaults to FALSE.}

\item{...}{Additional parameters passed on to the scale type}
}
\description{
Apply discrete color scale to ggplot object using eGRID palettes
}
