% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_capd_continuous}
\alias{scale_fill_capd_continuous}
\title{CAPD continuous fill scales}
\usage{
scale_fill_capd_continuous(palette = "sequential", reverse = FALSE, ...)
}
\arguments{
\item{palette}{name of palette, from \code{capd_palettes}. Defaults to "sequential".}

\item{reverse}{logical to indicate if palette order should be reversed. Defaults to FALSE.}

\item{...}{Additional parameters passed on to the scale type}
}
\description{
Apply continuous fill scale to ggplot object using CAPD palettes
}
