% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so2_aq.R
\docType{data}
\name{so2_aq}
\alias{so2_aq}
\title{SO2 Air Quality dataset}
\format{
A data frame with 129 rows and 4 columns
\describe{
\item{year}{year of data collection}
\item{units}{scientific units of SO2, ppb}
\item{stat}{label for SO2 summary statistic, either "avg" for average, "q10" for 10th \%ile, or "q90" for 90th \%ile}
\item{value}{value of SO2 summary statistic}
}
}
\source{
https://www.epa.gov/power-sector/progress-report-air-quality
}
\usage{
so2_aq
}
\description{
Collected from EPA Clean Air Power Sector 2022 Progress Report on Air Quality
}
\keyword{datasets}
