% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_capd.R
\name{theme_capd}
\alias{theme_capd}
\title{Set CAPD ggplot theme}
\usage{
theme_capd()
}
\value{
theme object to be appended to a ggplot call
}
\description{
Set CAPD ggplot theme
}
\examples{
# chart ggplot mpg dataset with CAPD theme
df <- ggplot2::mpg
ggplot2::ggplot(df, ggplot2::aes(x=cyl, y=cty)) +
  ggplot2::geom_point() +
  theme_capd()
}
