% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_egrid.R
\name{theme_egrid}
\alias{theme_egrid}
\title{Set eGRID ggplot theme}
\usage{
theme_egrid()
}
\value{
theme object to be appended to a ggplot call
}
\description{
Set eGRID ggplot theme
}
\examples{
# chart ggplot mpg dataset with eGRID theme
df <- ggplot2::mpg
ggplot2::ggplot(df, ggplot2::aes(x=cyl, y=cty)) +
  ggplot2::geom_point() +
  theme_egrid()
}
