% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capd_palettes.R
\docType{data}
\name{capd_palettes}
\alias{capd_palettes}
\title{Color Palettes for CAPD}
\format{
A named list with hex color codes
}
\source{
CAPD Data Visualization Style Guide
}
\usage{
capd_palettes
}
\description{
Color Palettes for CAPD
}
\keyword{datasets}
