% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egrid_palettes.R
\docType{data}
\name{egrid_palettes}
\alias{egrid_palettes}
\title{Color Palettes for eGRID}
\format{
A named list with hex color codes
}
\source{
eGRID Data Visualization Style Guide
}
\usage{
egrid_palettes
}
\description{
Color Palettes for eGRID
}
\keyword{datasets}
