% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_capd_discrete}
\alias{scale_fill_capd_discrete}
\title{CAPD discrete fill scales}
\usage{
scale_fill_capd_discrete(palette = "fuel_type", reverse = FALSE, ...)
}
\arguments{
\item{palette}{name of palette, from \code{capd_palettes}. Defaults to "fuel_type".}

\item{reverse}{logical to indicate if palette order should be reversed. Defaults to FALSE.}

\item{...}{Additional parameters passed on to the scale type}
}
\description{
Apply discrete fill scale to ggplot object using CAPD palettes
}
