% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_capd.R
\name{theme_capd}
\alias{theme_capd}
\title{Set CAPD ggplot theme}
\usage{
theme_capd(
  base_size = 11,
  plot_direction = "horiz",
  grid_lines = "major",
  ticks = FALSE,
  legend_position = "right"
)
}
\arguments{
\item{base_size}{font size, in pt}

\item{plot_direction}{direction to use for axis and grid lines. Options are "horiz", "vert", or "both".}

\item{grid_lines}{should major and/or minor gridlines be displayed? Options are "major","minor", or "both".}

\item{ticks}{should tick marks be displayed? Options are TRUE or FALSE.}

\item{legend_position}{side of chart to put legend on, use "right", "left", "top","bottom", or "none". passed to legend.position argument of theme}
}
\value{
theme object to be appended to a ggplot call
}
\description{
Set CAPD ggplot theme
}
\examples{
# chart ggplot mpg dataset with CAPD theme
library(ggplot2)
library(ggcapdthemes)
df <- resource_mix[resource_mix$year == 2023,]
p_bar <- ggplot(data = df) +
geom_bar(mapping = aes(x = generation_resource_mix, y = resource_label, fill = resource_label),
        stat = 'identity') +
 labs(x = 'Generation Resource Mix (\%)', y = 'Resource', fill = 'Resource',
      title = '2023 eGRID Generation Resource Mix',
      caption = 'Data from 2023 eGRID summary data, Table 2')
p_bar + theme_capd()
}
