% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_capd.R
\name{theme_capd}
\alias{theme_capd}
\title{Set CAPD ggplot theme}
\usage{
theme_capd(
  base_size = 11,
  axis_lines = "horiz",
  grid_lines = "major",
  ticks = FALSE,
  legend_position = "right"
)
}
\arguments{
\item{base_size}{font size, in pt}

\item{axis_lines}{orientation to use for axis and grid lines. Options are "horiz", "vert", or "both".}

\item{grid_lines}{should major and/or minor gridlines be displayed? Options are "major","minor", or c("major","minor").}

\item{ticks}{should tick marks be displayed? Options are TRUE or FALSE.}

\item{legend_position}{side of chart to put legend on, use "right", "left", "top","bottom", or "none". passed to legend.position argument of ggplot2::theme}
}
\value{
theme object to be appended to a ggplot call
}
\description{
Set CAPD ggplot theme
}
\examples{
# chart ggplot mpg dataset with CAPD theme
library(ggplot2)
library(ggcapdthemes)

# apply basic theme
gg_ex_bar + theme_capd()

# change base font size throughout plot elements
gg_ex_bar + theme_capd(base_size = 6)
gg_ex_bar + theme_capd(base_size = 16)

# rotate axis and grid lines
gg_ex_bar + theme_capd(axis_lines = "vert")

# show major and minor gridlines (default is major only)
gg_ex_bar + theme_capd(axis_lines = "vert", grid_lines = c("major","minor"))

# move legend to bottom of plot
gg_ex_bar + theme_capd(legend_position = "bottom")

## hide legend
gg_ex_bar + theme_capd(legend_position = "none")

## show ticks on non-major axis - depends on axis_lines value
gg_ex_bar + theme_capd(ticks = TRUE)
gg_ex_bar + theme_capd(axis_lines = "vert", ticks = TRUE)

}
