% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_capd.R
\name{theme_capd}
\alias{theme_capd}
\title{Set CAPD ggplot theme}
\usage{
theme_capd()
}
\value{
theme object to be appended to a ggplot call
}
\description{
Set CAPD ggplot theme
}
\examples{
# chart ggplot mpg dataset with CAPD theme
library(ggplot2)
df <- resource_mix[resource_mix$year == 2022,]
p_bar <- ggplot(data = df) +
geom_bar(mapping = aes(x = generation_resource_mix, y = resource_label, fill = resource_label),
        stat = 'identity') +
 labs(x = 'Generation Resource Mix (\%)', y = 'Resource', fill = 'Resource',
      title = '2022 eGRID Generation Resource Mix',
      caption = 'Data from 2022 eGRID summary data, Table 2')
p_bar + theme_capd()
}
